Fixes a minor spelling mistake when attempting to tweak access requirements on a vending machine (#9275)
# About the pull request

Fixes a minor spelling mistake when attempting to tweak access
requirements on a vending machine without the required skill.

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
spellcheck: Fixes a minor spelling mistake when attempting to tweak
access requirements on a vending machine
/:cl: