Adds a special sprite to locked flight controls (#9295)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->
Adds a icon to the flight navigation controls displaying when they're
locked.

# Explain why it's good for the game
It helps players better understand what is going on around them.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

For some reason, github isnt animating it, but it blinks

![dreammaker_1Rfh2f5Lhy](https://cdn.discordapp.com/attachments/745447048261795890/1373399288654331924/dreammaker_1Rfh2f5Lhy.gif?ex=682a4542&is=6828f3c2&hm=df421348f01afd164ba94d138e29a5d23f3ada66332f2282ba024682304f1687&)

![image](https://github.com/user-attachments/assets/a564925e-3241-40eb-a4cb-b3d05f499e37)


</details>


# Changelog
:cl:
imageadd: added a locked icon for locked dropship controls
/:cl: