RMC ert changes (#9124)
# About the pull request

Lowers RMC to 6, Increases recoil and decreases damage of L23, Changes
RMC Corpsman to more accurate Medical technician

# Explain why it's good for the game

RMC Amount: Lowered as the 3WE military is smaller then other
militaries.

L23: Lowered the damage to be more in line with other ERT guns, raised
to recoil as the current recoil doesn't feel enough to offset it's
current power.

Corpsman to Medical Technician: Corpsman is a purely American term.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
SRG3124

:cl:
balance: Lowered Maximum RMC ERT Personnel down to 6 instead of 7, 
balance: Ups recoil for the L23 and lowers the damage from 57.2 to 50.
/:cl:

---------

Co-authored-by: Detective-Google <48196179+Detective-Google@users.noreply.github.com>