Re-adds the (rebalanced) double barrel shotgun to the voteable maps it was removed from. (#9185)
# About the pull request

Self-explanatory, re-adds the Spearhead Rival 78 to maps where it was
removed in #3659 . I tried not to increase the amount of double barrels
that were in each map, and kept them the same types as they were
previously for the most part. My memory likely failed me in a spot or
two, but I put them in areas that made sense and areas I know for a fact
had one previously of the same type. I did not re-add them to any
nightmare inserts which previously included a double barrel, and instead
focused solely on the base maps, but I may do so in another PR down the
line.

I didn't re-add it to Sorokyne Strata since that map is getting a
massive revamp, and I think it's best not to touch it for now. LV-624
also didn't get the two double barrels re-added to the crashed ship in
containers, since I think it makes no sense for a supposed Wey-Yu ship,
with xeno specimens and PMC rifles, to be carrying around two double
barrels. I also hate Whiskey Outpost.

This is my first PR that changes regular maps and not nightmare inserts,
used the mapmerge tool before committing, hope there's no problems.

MAPS IT WAS RE-ADDED TO: Solaris Ridge, New Varadero, LV-624, Shiva's
Snowball, Kutjevo Refinery and Trijent Dam.

# Explain why it's good for the game

The double barrel was rebalanced a month or two ago by removing its
burst-fire feature and changing its fire delay to be on-par with the
MK221. It's now essentially a downgraded MK221 available on the colony.
It has cool sounds, it's an easily recognizable and iconic gun, and fits
perfectly for a ragtag bunch of colonists using anything and everything
they can to survive a horde of killing machines. It comes in very
limited quantities, too, as I didn't re-add it to the cargo black market
loot pool.

It also adds some nice variety to gunplay without breaking the game's
balance, you trade having to pause, divert your attention, and reload
more often for shooting faster and not having to pump your shotgun.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: IowaPotatoFarmer
add: Added the rebalanced double-barrel back to the maps it was removed
from, some in the same locations they were previously.
/:cl: