Armored gloves resprite (#9238)
# About the pull request

The current "armored gloves" sprite is just an ancient bay12 black
gloves sprite that looks ugly both onmob and as an item.
The glove is armored and insulated, but the sprite is just a classic
glove sprite, so i resprited it.
Inspired by the WY Elite glove sprites.

# Explain why it's good for the game

Ancient sprites dont look good / fit in very well.


# Testing Photographs and Procedure





![image](https://github.com/user-attachments/assets/e342a509-8dff-469f-905c-2bc573bba8f9)



# Changelog


:cl:Ansekishoku

imageadd: New sprites for armored gloves

/:cl:

---------

Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>