trijent LZs now powerable (#9259)
# About the pull request

the LZs on trijent had UNPOWERABLE set to 1 which caused confusion as
this was a departure from other maps where the expectation is that the
marine FOB can be powered.

# Explain why it's good for the game

the marine FOB being able to be powered is standard, and is part of the
incentive for controlling power generation as the marine faction,
arbitrarily disallowing power from applying to this area is confusing.




# Changelog
:cl:
maptweak: trijent LZs can now be powered
/:cl: