Xenos now heal at normal speed when not on fire (#9364)
# About the pull request

Current healing checks if there are any fire stacks currently on the
xeno, not if the xeno is actually on fire.
Xenos with fire stacks while not on fire is basically a rag soaked in
petrol, very flammable but not harmed.

Fixes: #9363 and xeno ignition immunity boon healing slowly while not on
fire.

# Explain why it's good for the game

Xenos not on fire should not have healing penalized for flame stacks
that have yet to be lit on fire.

# Testing Photographs and Procedure

None.


# Changelog

:cl: KornFlaks
fix: Xenos now heal at full speed while not on fire.
/:cl: