prevents mar-50 from attaching a bipod (#9323)
# About the pull request

removes the bipod from the list of valid attachments for the mar-50 LMG
fixes minor bug where the mar-50 could spawn with invalid attachments

# Explain why it's good for the game

same reasoning as #8768

The mar-50 with a bipod is too powerful right now, and can obliterate
heavily armored t3s in about a second from full screen range, which is
too much. My sincerest apologies to the survcord.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: the mar-50 LMG can no longer be used with a bipod
fix: the mar-50 no longer spawns with other invalid attachments like
barrel attachments
/:cl: