fixes specs in crew monitor (#9348)
# About the pull request

As title

# Explain why it's good for the game

Good idea not to have specs and supporting roles at the bottom of the
crew monitor.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: SHARP Operator, Heavy Sniper, and marine specialist support roles
now show properly on the crew monitor.
/:cl:

Co-authored-by: forest2001 <forest2001@cm-ss13.com>