Changes the Jungle Moon hunting grounds map (#9292)
# About the pull request

- Updates the Jungle Moon hunting ground map to be closer in and
claustrophobic to dissuade people from rushing to certain areas of the
map on spawn.
- The Jungle Moon now has its own ambience and animal sounds, thank you
to the person who provided the animal sounds, you know who you are.
- Areas now have a hangable trait to allow for Yautja to hang up bodies
where there might not be a roof, but where it would be feasible for
bodies to be hung from trees IE. Jungle tiles.
- Fixes some upside-down tiles on LV-624.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
The current map for the jungle moon is not bad. But it is the opinion of
many people that it encourages hunted individuals to rush to certain
areas of the map. Mainly, the open ground next to the camp. This version
of the map is more claustrophobic, with fewer open areas for people to
rush towards on spawn.

Fixing upside-down tiles on LV is long overdue.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

### See the map diff bot for the map. These are OGGs to MP4s. I am aware
they are black screens.


https://github.com/user-attachments/assets/387e53f0-a5d1-4d78-9415-0e0ff2b85a40



https://github.com/user-attachments/assets/2ee04d47-2784-4841-9fb4-0fde804246fe


https://github.com/user-attachments/assets/8cb9bc3a-543a-448b-9ea9-c13b2ce002d1



https://github.com/user-attachments/assets/4344b940-049d-47f5-a923-43b2ff9c06a5



</details>


# Changelog
:cl: Joe Lampost, Suprisedbaby_
add: Areas now have a hangable var that allows for the hanging of bodies
where it would be feasible for them to be hung.
fix: Fixes some upside-down tiles on LV-624.
soundadd: Added New ambience for the Jungle Moon.
soundadd: Added three new random animal noises for the Jungle Moon.
maptweak: Updated the Jungle moon.
/:cl: