Adds supply radio to SL headsets (#9336)
# About the pull request

Adds supply radio to SL headsets. Fixes descriptions on headsets to fit
this new change.

# Explain why it's good for the game

QoL change that helps SL's be able to properly talk to Req outside of
having to play literal or figuratively telephone through radio backpack
/ asking FTL or SO to talk to req for you.

Since SL's that know what they are doing tend to need to talk to req a
decent bit, and FTL's already have this channel, I do not see the harm
of giving it to SL's as well.

Yes you can buy supply radio key or ask req, but it should be default
considering not having good comm channels like supply radio as a SL is
like not having medkits as a corpsmen or not having your tools for free
as a engineer.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

N/A

</details>

# Changelog

:cl: Swagile
add: Gives SL radio supply channel access by default.
/:cl: