Allows Yautja to Flay Prey with the Ceremonial Dagger Instead of Butcher Verb (#9243)
# About the pull request

Allows Yautja to flay a target with the Ceremonial Dagger and have it
unset the current target (like the `butcher` proc does).

Tested it and ensured that `butcher` still functions correctly and is
unchanged. Time-to-flay is approximately the same as the `butcher` verb
so unless I need to change it to be exactly equal I'm leaving it as-is.

The proc only tracks the target trophy being taken when the scalp is
removed.

# Explain why it's good for the game

During my first Youngblood round, I was unaware there even **was** a
`butcher` verb. So, logic dictated that you use your bootknife (the
Ceremonial Dagger) to skin downed Prey.

This, apparently, doesnt do anything. Well now it does.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: flaying a target with the Ceremonial Dagger now announces and
unsets Prey like the Butcher Verb.
/:cl: