Nuke messaging clarity (again) (#9418)
# About the pull request

Currently, halfway through the detonation countdown is referred to as a
'preparation cycle' in xeno messaging. Changes messaging to 'detonation
countdown'. Continuation of
https://github.com/cmss13-devs/cmss13/pull/7592

# Explain why it's good for the game

"Preparation cycle" is confusing, even for me as a bazillion hour
player. Isn't the preparation cycle the decryption phase? Can we call
the detonation countdown "the detonation countdown"?


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog

:cl:
qol: nuke detonation countdown is now called detonation countdown, not
"preparation cycle"
/:cl: