One retry on relay browser (#9426)
# About the pull request

This PR is a follow up to #5777 adding one retry on error.

It won't necessarily be in ping order, but the order is based on the
original response time for the first attempt.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

I faked this failure because it doesn't seem the relays are currently
overloaded but should look like this:

![retry](https://github.com/user-attachments/assets/65612517-7509-4f10-b42c-40964b3b52aa)

</details>


# Changelog
:cl: Drathek
ui: Ping relay browser will attempt to retry once if a relay errors
/:cl: