trashcart infinte money glitch patch (#9435)
# About the pull request
swaps the trash cart (in the context of this one black market instance!)
into an already existing subtyped crate so that you can't sell it to
normal ASRS for infinte money.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

infinte money bad

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog


:cl:

fix: infinite money glitch
/:cl: