Youngblood Hour changes. (#9342)
# About the pull request
This PR reduces the required hours for Youngblood while also adding a
hard cap on the hours people can play Youngblood for.
Whilst also removing the lock on one youngblood ERT per round. Instead
adding a 40 minute cooldown.

Three new ERTs have been added related to youngblood:

- Inexperienced
The hours required for this ERT are 5 hours, as xeno and marine. This
ERT is intended for players who have never touched Youngblood before.
Players who have more than two hours as youngblood will not be able to
roll for this ERT.

- Intermediate
The hours required for this ERT are 10 hours. This ERT is intended for
players who are more familiar with the mechanics of the game. Players
who have more than five hours as Youngblood will not be able to roll for
this ERT.

- Experienced
The hours required for this ERT are 20 hours. This ERT is intended for
players who are more experienced. Players who have more than 10 hours as
Youngblood will not be able to roll for this ERT.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

The Youngblood ERT is intended to give players a taste of the Yautja WL,
hopefully incentivising them to apply. Unfortunately, some players have
abused this privilege by repeatedly taking the role to play as
Youngblood as often as possible without actually applying. While
reducing the time required for the role broadens the pool of players who
can try a new faction, which is always positive.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds a hard lockout for player who have reached 10 youngblood hours
from youngblood.
add: Adds a cooldown for youngblood ERTs (40 minutes).
add: Adds three new youngblood ERTs intended for newer and more
experienced players
/:cl: