Adds button to toggle displaying Plasma on the updated Hive Status (#9441)
# About the pull request

Plasma level is somewhat relevant for Queens, and not at all relevant
for most other castes.

Adds a button in the UI to toggle the column. Default option is ON for
the Queen when on Ovi, and OFF for anyone else.
(It's only really a valid piece of info when you have the 'Transfer
Plasma' button next to it. Can be changed if we want)

# Explain why it's good for the game

Less clutter, yay!

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/2a9b92e2-f290-46a5-a98b-c74f7661e96d)


![image](https://github.com/user-attachments/assets/e228e16e-541e-4852-959c-444b404cb3e0)


</details>


# Changelog

:cl:
ui: There's now a 'Show Plasma' toggle in the Hive Status UI to toggle
the Plasma column.
/:cl: