Weyland-Yutani Intranet Terminal (#7511)
# About the pull request

Adds a console for navigating the Weyland-Yutani Intranet. As of the
moment it's quite barebones, but can be used by the Liaison to control
their hidden containment cell.

# Explain why it's good for the game

Makes the containment cell controls more secure, and gives the CL's desk
a functional computer.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a Weyland-Yutani Intranet Terminal to the CLs office.
maptweak: Re-orders the buttons on the CLs desk to make more sense.
/:cl:

---------

Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>
Co-authored-by: forest2001 <forest2001@cm-ss13.com>