Cigarette Lore (#9403)
# About the pull request

This PR adds extended descriptions to the three most common used
cigarette brands in the game.


# Explain why it's good for the game

Cigarettes are some of the very first items marines interact with when
waking up from hypersleep. I feel that by adding fluff like this, it
encourages new and old players alike to invest themselves more into the
setting, and maybe roleplay a little more.

Additionally I've noticed that `desc_lore` is still _mostly_ being used
for guns which is a shame honestly because it can add so much
storytelling and flavor to what would otherwise be a boring and plain
experience. Brand loyalty isn't just about appearances.


# Testing Photographs and Procedure


![image](https://github.com/user-attachments/assets/c0654bcf-8589-4779-9c5d-afa6eba480e0)


# Changelog
:cl:
add: Adds extended lore to three brands of cigarettes.
/:cl: