Helps new players understand hivemind collapse. (#9567)
# About the pull request

Adds text telling players that a queen needs to evolve or the hive
cannot survive much longer.

# Explain why it's good for the game

The game ending due to hivemind collapse (no queen for a certain amount
of time) is a very strange concept for many players, especially newbies.
We might as well warn players that a queen needs to evolve soon or the
hive won't survive.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
qol: give hint regarding impending game end due to hivemind collapse
/:cl: