Another labeler fix (#9561)
# About the pull request

This PR is a follow up to #9428 to add a couple more needed usages of
has_label. Basically since we rarely ever qdel the label component, when
a label is cleared the component still knows what the old label was and
this was preventing a cleared label to block a new label with the same
name.

# Explain why it's good for the game
Fixes #9556 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/sSWyEMwfyoI

</details>


# Changelog
:cl: Drathek
fix: Fixed a cleared label preventing new a label if it was the same as
the last label
/:cl: