acid now moves with acided things, you can no longer pickup acided items (#9256)
# About the pull request

When acid is applied to an item, you can't pick it up until the acid
goes away. (you can still drag it, or pull things out of it if it is a
bag etc.)

Acid now moves with the item that was acided rather than being anchored
in place where it was applied.

# Explain why it's good for the game
#9241 seeks to make spraying the acid off of a weapon part of the
gameplay loop, for this to be practical, the acid needs to stay with the
weapon as a visual indicator that it is melting.

Acid being left behind when objects are removed causes undue frustration
and confusion for new players, who may have their weapon returned to
them having had acid applied, only for it to be deleted from their hands
with little warning. This rectifies that.

It also just doesn't make that much sense to be putting your hands on
something covered in acid and touching it.

# Testing Photographs and Procedure
<!-- Include any screenshots/videos/debugging steps of the modified code
functioning successfully, ideally including edge cases. -->

<!-- !! If you are modifying sprites, you **must** include one or more
in-game screenshots or videos of the new sprites. !! -->

<details>
<summary>Screenshots & Videos</summary>




https://github.com/user-attachments/assets/d5a3366c-2973-4682-8f9b-0503a0d69297




</details>


# Changelog

:cl:
add: acid now tracks the item it is applied to
balance: you can't pick up items with acid on them
/:cl:

<!-- Both :cl:'s are required for the changelog to work! -->

---------

Co-authored-by: fira <loyauflorian@gmail.com>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>