Reverts the jamming system, replaces it with a single saved acid attack. (#9241)
# About the pull request
Reverts the jamming system #8682 , and the peripheral PRs afterwards
#9030 , #9074 , #9073 , #9028

Guns now have a "Second wind" that can be seen in their examine.

Spraying an acided gun with an extinguisher will extinguish the acid.

Allowing the acid to finish on the weapon will remove it's "Second wind"
and play a click noise.

Once the weapon has lost its "Second wind" it can no longer have future
acid sprayed off of it.

"Second wind" applies to explosions from minor to moderate severity, and
is only consumed if the weapon would have otherwise been destroyed by
the explosion.

# Explain why it's good for the game

The jamming PR sought to remove some of the unfun gameplay elements of
having your weapon destroyed because a xeno stood next to it for ~1
second, being irrecoverably trashed. It intended to add depth to the
marine side by making your gun something you had to take care of beyond
an arbitrary "works perfectly" or "is qdeleted" The general goal was to
add a tense interruption to gameplay as a punishment for not taking care
of your weapon that was dramatic, but not tedious.

I still believe those are good goals to look after, and that the PR was
a good attempt at solving them, but I also believe it had some issues
that are so fundamental as to demand a full rewrite before
reimplementation.

1. The change made weapons too difficult to attrition

Weapon attrition is a big part of xeno attrition and one of the tasks
acid carrying castes perform on the front line, the lowered marine
weapon attrition reduced the value of these already somewhat weak
castes, and made replacement weapons/attachments/etc a somewhat
nonissue.

2. The change made dropping your weapon too inconsequential. 

While it is definitely un-fun to lose your weapon because you were
knocked away from it for a moment, especially considering the prevalence
of stuns and flings, the current iteration makes it too easy to not care
about your weapon, this meant you could run with your weapon ready at
all times, and without a harness, or anywhere to holster it, with very
minimal concern as to whether it will be destroyed.

3. The jamming mechanic was confusing to players. 

The process of maintaining a weapon, what could damage a weapon, why a
weapon would jam and at what durability was confusing and frustrating to
players, examining a weapon and being told it was "fine" durability and
then having it jam and get the user killed was intensely bothersome.
Un-jamming a weapon was similarly unsatisfying, due to the RNG elements,
which meant that sometimes it could take an unfair amount of attempts to
un-jam a weapon before it would fire again. Rather than creating a panic
in the tempo of combat it just created frustration or confusion at lost
opportunity for players.

4. The durability system was obtuse and un-engaging to interact with for
players.

It was not always clear when a weapon had become damaged, why, or to
what extent. A weapon that read as "fine" on its description could be
nigh unusable depending on a combination of other factors/variables
hidden to the user. Due to inconsistent implementation this affected
some weapons much more harshly than others, which hardly ever even saw
jams and barely interacted with the system, leading to a perception of
"jank" Similarly, repairing a weapon wasn't engaging and ended up being
a matter of standing still for ~20 seconds, even with the cooperation of
a teammate with engineering skill, which was intensely uninteresting and
didn't add to the feeling of maintaining a weapon. The quicker, lesser
option in gun oil was so ineffective as to be similarly unsatisfying, or
frustrating when after application the weapon still did not function.

5. The code for the system is substandard and hard to read. 

The code for the jamming system was based on the work of another coder
from a time with different standards and tools available, and was a
learning project both for the PR creator and myself as a reviewer. In
the process of testing, reviewing, rereading, and suggesting changes to
the PR I became somewhat blind to some of the issues in structure it
had. While the code is mostly up to the literal standards of the
codebase, it is implemented in a manner that makes maintaining it and
changing it highly unintuitive, to the point it warrants a reversion,
and a rewrite to use signals and the datum component system when/if
re-implemented.


<details>
<summary>Screenshots & Videos</summary>

</details>


# Changelog

:cl:
del: Reverts the jamming system
add: Guns now have a "Second wind" that can be seen in their examine.
add: Spraying an acided gun with an extinguisher will extinguish the
acid.
balance: Allowing the acid to finish on the weapon will remove it's
"Second wind" and play a click noise.
balance: Once the weapon has lost its "Second wind" it can no longer
have future acid sprayed off of it.
add: "Second wind" applies to explosions from minor to moderate
severity, and is only consumed if the weapon would have otherwise been
destroyed by the explosion.
/:cl:

- [x] todo: add extinguisher req to the second wind
- [x] todo: save can prevent minor explosions from destroying wep