tweak: Xeno can be fed with more types of food (#9545)
# About the pull request
Xenomorphs can be fed with any /obj/item/reagent_container/food/snacks,
so why not remove the restriction on the ability to consume other types
of food/drink for xenomorphs?
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
**Xeno can drink beer! Its cool!**
# Testing Photographs and Procedure
below
<details>
<summary>Screenshots & Videos</summary>


https://github.com/user-attachments/assets/addd1724-c8d9-490e-91d5-1e9d5853e319


</details>


# Changelog
:cl:
add: Xeno can be fed with any type of food
/:cl:

Co-authored-by: LiquidPotroh <max.dobro123@gmail.com>