Enables the storage of flags and tesla into the construction belt. (#9425)
# About the pull request

Enables the storage of flags and teslas into the construction rig belt.

I do believe it might be an oversight, considering all sentries, flame
sentries and their "upgraded" counterparts are able to be put inside
them.

# Explain why it's good for the game

Puts the tesla and flag in line with the other two sentries when it
comes to storage and prevents attempts at putting the flags/teslas into
the construction rig thinking it could be done the same as the
flamer/normal sentries, possibly wasting their belt choice.


# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>

![Capture d'écran 2025-06-02
140712](https://github.com/user-attachments/assets/4d7384cc-05e6-4431-a51b-19787f8bee73)

</details>


# Changelog
:cl:
balance: Enabled the storage of flag/teslas into the construction rig.
/:cl: