Rename `var/mob/usr` to `var/mob/user` (#9592)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

OpenDream is adding `usr` to the `SoftReservedKeyword` pragma:
https://github.com/OpenDreamProject/OpenDream/pull/2307

This PR renames `var/mob/usr` to `var/mob/user`.

# Explain why it's good for the game
This pragma is currently set as an error in CM, because using internal
var names for declared var names is cringe.


# Changelog
it's a var rename