Add Antag Skill to CIA Agent + Amend a variable for their portable vendor (#9751)
# About the pull request
Adds Antag skill at Agent level so CIA Agents can use the chloroform
they can buy.
Moves some seconds that should be in the var to the var
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Bugfix + Uhhhhhhh Code Style Standards or something
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: CIA Skills include Antag for the items they can buy that require
it.
/:cl: