No larva queue estimation for the first 15 seconds from lobby (#9773)
# About the pull request

This PR is a follow up to #7261 simply preventing the estimation of a
position number when you're on the lobby for the first 15 seconds of the
game.

# Explain why it's good for the game

Larva queue information endlessly confuses players.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/d79cecef-0505-4d0e-bfcd-625fa99ca2e4)

</details>


# Changelog
:cl: Drathek
qol: First 15 seconds of the game no longer offers a larva queue
position estimation on the lobby since other players are still observing
and the estimation relies on currently valid candidates.
/:cl: