Tweaks the mortars ranges (#9549)
# About the pull request

Title.

# Explain why it's good for the game

Currently, the mortar is a bit intrusive to use, even especially on the
front due to its minimum range requirements since it requires three and
roughly a quarter screens to use properly meaning if you are the sole
user, you are forced to be at constant pace back and forth relazing and
recalculating on the go.

I personally think that lowering it to at least 2 screens worth still
allows for more legroom, while also not a pain to utilize during FOB
sieges, with the constant repositioning that it currently requires for
it.

Max range is also lowered to compensate

# Testing Photographs and Procedure
Line change.

# Changelog

:cl:
balance: Mortar minimum range is lowered from 25 to 15, maximum range
from 75 to 64
/:cl: