M6H-BRUTE (#9310)
# About the pull request
Adds the M6H-BRUTE Launcher System to the ComTech Equipment vendor as an
alternative option to sentry guns for the more offensive engineer. It
launches laser guided rockets at turfs and structures and destroys a
swath of walls and such in a sufficiently large area, leaving behind a
wake of shrapnel and fire. It does not do significant damage to
Xenomorphs, and is requires a laser-guided aiming sequence before
firing.

- Spawns in ComTech Equipment Vendor, as an option alongside the Sentry
Guns, which it can be taken as an alternative to.

- Comes with the Launcher (loaded), 2 spare Rockets, all stored in the
Rig.

- The Rig itself works like a holster. Worn on the back, the rig can
store the launcher and up to 6 additional slots of ammunition. The slots
can hold both rockets and regular ammunition types (mags, shells, etc.)

- The launcher is a large item that has to be wielded in order to be
fired. To fire, user must laser target a structure or wall (similar to
sniper's aimed shot), and after a brief moment of locking on, the
launcher will be able to fire, after which it will travel directly to
the target and explode on impact.

- The rocket, upon exploding, will deal heavy damage to structures in a
7x4 tile ellipse offset from the point of impact, directionally. This
explosion should not do much damage to mobs, and at most should fling
mobs away from point of impact, leaving some very weak welder fuel fire
sporadically on the ground in its wake.

- Similar to the M5 RPG, the launcher can then be reloaded with spare
rockets. Additional rockets can be purchased at Requisitions in batches
of 2.

- ComTech only, should require a degree of both weapons and engineering
skills to use.


# Explain why it's good for the game

It provides Marines good breaching utility at a range without being
infinite ammo, and requires a little thought and coordination while
using, as opposed to mindlessly deleting entire buildings with the XM51.
Plus, explosions are badass.

# Testing Photographs and Procedure


![qrbxi70AJg](https://github.com/user-attachments/assets/844f9f34-0535-4430-a55e-9c7b3a3210f8)

![dreamseeker_cBKZQyHLOn](https://github.com/user-attachments/assets/7f6e8a55-6bca-4857-b524-524c593705fe)


<details>
<summary>Screenshots & Videos</summary>

https://i.imgur.com/wHMzEkV.gif

</details>


# Changelog
:cl: Thwomper, cuberound
add: M6H-BRUTE, a anti-structure laser-guided rocket launcher, as well
as a specialized back holster and ammo to ComTech as an alternative
equipment choice to sentry guns.
del: XM-51 Breaching Scattergun no longer available in Marine supply. 
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>
Co-authored-by: cuberound <122645057+cuberound@users.noreply.github.com>
Co-authored-by: Git-Nivrak <59925169+Git-Nivrak@users.noreply.github.com>
Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>