Pylon-Layer-Changes (#9761)
# About the pull request

Changes the layer of Pylons to be below things like lockers and doors. 

https://forum.cm-ss13.com/t/please-change-the-pylon-layer/14972

# Explain why it's good for the game

Hiding a structure so important to the game behind a locker is a lil
cheesy.

# Testing Photographs and Procedure

<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/76f034f6-c5fa-4cb9-bede-ac783d2a7691)

Pylon without the layer changes on the left, pylon with the layer
changes on the right. Looks a lil jank so if a spriter wants to extend
the base a bit I won't complain!

</details>


# Changelog

:cl:
balance: Changes the pylon layer, preventing it from being hidden behind
stuff like lockers.
/:cl: