Tiny fix for RMC ERT helmet slot. (#9841)
Fixes a mistake from a previous PR that had overwritten a RMC helmet for
the spec.
# About the pull request

Tiny mistake from a previous PR that accidently replaced the RMC specs
helmet with an IASF beret, it should be
/obj/item/clothing/head/helmet/marine/veteran/royal_marine rather then
the IASF beret.

# Explain why it's good for the game

Unintended mistake, should be reverted.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Zenith

fix: Fixes a mistake from a previous PR that had overwritten a RMC
helmet for the spec.

/:cl: